refactor(uikit): move address related components and data structures to proper packages
