+++ examples/shop-with-blog/client/package.json
@@ -29,6 +29,7 @@
     "@deity/falcon-shop-data": "^0.0.1",
     "@deity/falcon-theme-editor": "^0.1.5",
     "@deity/falcon-ui": "^0.2.2",
+    "@deity/falcon-uikit": "0.0.1",
     "@loadable/component": "5.7.0",
     "apollo-client": "2.5.1",
     "config": "^2.0.1",
