+++ examples/shop-with-blog/client/src/pages/shop/Account/AddressBook/EditAddress.js
@@ -8,15 +8,13 @@ import {
   FormField,
   CheckboxFormField,
   FormErrorSummary,
-  AddressQuery,
-  getAddressType,
-  GET_ADDRESS,
-  EditAddressMutation,
   TwoColumnsLayout,
   TwoColumnsLayoutArea,
   CountriesQuery,
   CountrySelector
 } from '@deity/falcon-ecommerce-uikit';
+import { getAddressType } from '@deity/falcon-ecommerce-uikit';
+import { GET_ADDRESS, AddressQuery, EditAddressMutation } from '@deity/falcon-shop-data';
 
 const EditAddress = ({ match, history }) => {
   const id = parseInt(match.params.id, 10);
@@ -34,7 +32,7 @@ const EditAddress = ({ match, history }) => {
               <T id="editAddress.defaultAddressLabel" context={getAddressType(address)} />
             </Text>
           )}
-          <EditAddressMutation refetchQueries={['Addresses', { query: GET_ADDRESS, variables: { id } }]}>
+          <EditAddressMutation refetchQueries={['AddressList', { query: GET_ADDRESS, variables: { id } }]}>
             {(editAddress, { loading, error }) => (
               <Formik
                 initialValues={{
