--- packages/falcon-ecommerce-uikit/src/Address/index.tsx
@@ -1,3 +1,0 @@
-export * from './Address';
-export * from './AddressQuery';
-export * from './AddressMutation';
