--- packages/falcon-ecommerce-uikit/src/Customer/ChangePasswordMutation.tsx
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag';
-import { Mutation } from 'react-apollo';
-
-export const CHANGE_PASSWORD = gql`
-  mutation changePassword($input: CustomerPassword!) {
-    changeCustomerPassword(input: $input)
-  }
-`;
-export class ChangePasswordMutation extends Mutation {
-  static defaultProps = {
-    mutation: CHANGE_PASSWORD
-  };
-}
