--- packages/falcon-ecommerce-uikit/src/Customer/index.ts
@@ -1,4 +1,0 @@
-export * from './CustomerQuery';
-export * from './CustomerMutation';
-export * from './ChangePasswordMutation';
-export * from './IsAuthenticatedQuery';
