+++ packages/falcon-shop-data/src/Address/AddressQuery.tsx
@@ -1,0 +1,28 @@
+import gql from 'graphql-tag';
+import { Address } from '@deity/falcon-shop-extension';
+import { Query } from '../Query';
+
+export const GET_ADDRESS = gql`
+  query Address($id: Int!) {
+    address(id: $id) {
+      id
+      firstname
+      lastname
+      telephone
+      street
+      city
+      postcode
+      region
+      regionId
+      countryId
+      company
+      defaultBilling
+      defaultShipping
+    }
+  }
+`;
+export class AddressQuery extends Query<Address> {
+  static defaultProps = {
+    query: GET_ADDRESS
+  };
+}
