+++ packages/falcon-uikit/src/Address/index.ts
@@ -1,0 +1,4 @@
+export * from './AddressCard';
+export * from './AddressDetails';
+export * from './AddressListLayout';
+export * from './utils';
