+++ packages/falcon-uikit/src/Address/utils.ts
@@ -1,0 +1,6 @@
+import { Address } from '@deity/falcon-shop-extension';
+
+export type AddressType = 'billing' | 'shipping' | 'billing&shipping' | 'other';
+export const getAddressType = (address: Address): AddressType =>
+  ([address.defaultBilling && 'billing', address.defaultShipping && 'shipping'].filter(x => x).join('&') ||
+    'other') as AddressType;
